fix id in entities

+++ backend/src/main/java/com/netcraker/edu/backend/entities/BaseEntity.java
@@ -9,20 +9,20 @@ public class BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false, updatable = false)
-    private int id;
+    private long id;
 
     public BaseEntity() {
     }
 
-    public BaseEntity(int id) {
+    public BaseEntity(long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
